Added yield(sec) function with simple polling.
Added a function yield(sec) that will wait the number of seconds
before continuing with the message. Simple polling via the
avatar will return result if time is up and message continues to
run or false if time isn't up yet.

(Also a tiny bug fix for when you set a global = 0)