Add system.wait() and support in client.js
The tricky thing is that it works by inserting a string into
the compiled message, and then the client needs to detect
the string and do the wait. This means that you have to use
the <%= %> when using wait() so that the string is inserted
by the template engine.

Okay, I should probably mention client.js and its sleep method.
It's awful, it sucks, it's terrible and should never be done in
nodejs server code. Luckily, it will never run on a server because
it's just for local development support. Whew! But seriously kids,
don't block your node process!